Refactoring to match common code styles
