+++ src/main/java/game_engine/controller/utility/Utils.java
@@ -3,11 +3,11 @@ package game_engine.controller.utility;
 import java.util.UUID;
 
 /**
- * Included useful functions which are needed in multiple class instances
+ * Included useful functions which are needed in multiple class instances.
+ *
  * @author FabianDev001
  */
 public class Utils {
-
 
     public Integer generateNewRandomID(){
         return UUID.randomUUID().hashCode();
@@ -17,7 +17,8 @@ public class Utils {
      * An addition to the cmd-logger - In the future will create a log file for details
      * @param message
      */
-    public void log_red(String message){
+    public void logRed(String message){
         System.out.println("\u001B[31m" + message + "\u001B[0m");
     }
+
 }
